requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-bootstrap==0.2.4
django-bootstrap4==21.1
django-bootstrap5==21.2
django-haystack==2.4.0
django-rest-framework==0.1.0
django-taggit==2.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.3
PyJWT==1.6.4
pyparsing==3.0.7
pysolr==3.3.2
pytz==2018.5
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
